requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

aiokafka
sqlalchemy

redis

websockets

tenacity

loguru

pytest
pytest-asyncio
